refactor(core): move Claude subagent prompt builder into fleet-cli hook
The `<fleet section="subagents">` SessionStart additionalContext is a
host event hook artifact, which doctrine assigns exclusively to the
fleet-cli side. fleet-admiral never consumed `buildSubagentsSection`
internally, so it does not belong on the Admiral public surface.

Move `buildSubagentsSection` and `SubagentSectionEntry` into a new
`runtime/fleet-cli/src/hooks/subagents-context.ts` module alongside the
hook logic, so fleet-cli reads persisted carrier state via the
fleet-carriers store and builds the prompt itself. Drop the symbols from
the fleet-admiral barrel, public-surface whitelist, and tests, and prune
the now-dead imports from the CLI entrypoint. Generated prompt output is
byte-identical.
