+++ packages/fleet-admiral/src/index.ts
@@ -1,7 +1,5 @@
 export {
-  buildSubagentsSection,
   createSystemPromptBuilder,
-  type SubagentSectionEntry,
   type SystemPromptBuilder,
 } from "./prompts.js";
 export {
