+++ runtime/fleet-cli/src/hooks/subagents-context.ts
@@ -1,0 +1,95 @@
+import { readFileSync } from "node:fs";
+import os from "node:os";
+import path from "node:path";
+
+import {
+  buildClaudeSubagentDefinitions,
+  createCarrierRuntime,
+  getCarrierConfig,
+  getEnabledCarrierSubagentIds,
+  getRegisteredOrder,
+  readCarrierAgentModeSnapshot,
+  resolveAgentCliType,
+  type CarrierConfig,
+  type CarrierModelDefaults,
+} from "@dotobokuri/fleet-carriers";
+
+interface SubagentSectionEntry {
+  readonly carrierId: string;
+  readonly displayName?: string;
+  readonly nativeName: string;
+}
+
+export function runSubagentsContextHook(env: NodeJS.ProcessEnv): string {
+  const fleetRoot = env.FLEET_ROOT ?? path.join(env.HOME ?? os.homedir(), ".fleet");
+  if (!canReadCarrierState(path.join(fleetRoot, "carriers.json"))) {
+    return JSON.stringify({ hookSpecificOutput: { hookEventName: "SessionStart", additionalContext: "" } });
+  }
+  const carrierRuntime = createCarrierRuntime();
+  carrierRuntime.store.initStore(fleetRoot);
+  carrierRuntime.registerCarrierDefaults();
+  const carrierIds = getRegisteredOrder(carrierRuntime.registry);
+  const carrierConfigs = carrierIds
+    .map((carrierId) => getCarrierConfig(carrierRuntime.registry, carrierId))
+    .filter((config): config is NonNullable<typeof config> => config !== undefined);
+  const defaultsByCarrier = Object.fromEntries(
+    carrierConfigs.map((config) => [config.id, buildCarrierModelDefaults(config)]),
+  );
+  const enabledCarrierIds = getEnabledCarrierSubagentIds(
+    readCarrierAgentModeSnapshot(defaultsByCarrier),
+    carrierIds,
+  );
+  const definitions = buildClaudeSubagentDefinitions({ carrierConfigs, enabledCarrierIds });
+  const configsById = new Map(carrierConfigs.map((config) => [config.id, config]));
+  const additionalContext = buildSubagentsSection(definitions.map((definition) => ({
+    carrierId: definition.carrierId,
+    displayName: configsById.get(definition.carrierId)?.displayName,
+    nativeName: definition.name,
+  }))) ?? "";
+  return JSON.stringify({ hookSpecificOutput: { hookEventName: "SessionStart", additionalContext } });
+}
+
+function buildSubagentsSection(entries: readonly SubagentSectionEntry[]): string | undefined {
+  if (entries.length === 0) return undefined;
+  const lines = entries
+    .map((entry) => {
+      const label = entry.displayName ? `${entry.displayName} (${entry.carrierId})` : entry.carrierId;
+      return `- ${label}: invoke as Claude native subagent \`${entry.nativeName}\`.`;
+    })
+    .join("\n");
+  return `<fleet section="subagents">\n# Claude Native Subagents\n\nThe following Fleet carriers are exposed as Claude native subagents for this session:\n\n${lines}\n\nNative subagent calls return inline and do not emit \`[carrier:result]\`. Do not wait for a carrier job completion push after native invocation.\n\n\`carrier_dispatch\` remains available as a separate Fleet delegation path for carriers that are not invoked through the native subagent interface.\n</fleet>`;
+}
+
+function canReadCarrierState(filePath: string): boolean {
+  try {
+    return isReadableCarrierStateRoot(JSON.parse(readFileSync(filePath, "utf8")));
+  } catch {
+    return false;
+  }
+}
+
+function isReadableCarrierStateRoot(value: unknown): boolean {
+  if (!isRecord(value)) return false;
+  const carriers = value.carriers;
+  return carriers === undefined || isRecord(carriers);
+}
+
+function isRecord(value: unknown): value is Record<string, unknown> {
+  return typeof value === "object" && value !== null && !Array.isArray(value);
+}
+
+function buildCarrierModelDefaults(config: CarrierConfig): CarrierModelDefaults {
+  const cliType = resolveAgentCliType(config.id, config.defaultCliType);
+  const cliDefaults = cliType === "claude"
+    ? config.subagent?.byHost?.claude ?? {
+      ...(config.defaultEffort ? { defaultEffort: config.defaultEffort } : {}),
+      ...(config.defaultModel ? { defaultModel: config.defaultModel } : {}),
+    }
+    : {};
+  return {
+    cliType,
+    ...(config.defaultAgentMode ? { defaultAgentMode: config.defaultAgentMode } : {}),
+    ...(cliDefaults.defaultEffort ? { defaultEffort: cliDefaults.defaultEffort } : {}),
+    ...(cliDefaults.defaultModel ? { defaultModel: cliDefaults.defaultModel } : {}),
+  };
+}
